Reduce duplicate code in wrappers for os
